requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.23
djangorestframework==3.15.2
et-xmlfile==2.0.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2